Rydder i headerkode( en komponent i hver fil)
