+++ src/view/dialog/DialogHeader.module.less
@@ -31,10 +31,6 @@
     overflow: hidden;
 }
 
-.dialogHeaderContainer {
-    display: block;
-}
-
 .dialogHeader {
     display: flex;
     align-items: center;
