+++ src/view/dialog/DialogHeader.tsx
@@ -1,16 +1,15 @@
 import { hasError } from '@nutgaard/use-fetch';
 import classNames from 'classnames';
 import { VenstreChevron } from 'nav-frontend-chevron';
-import { Systemtittel, Undertittel } from 'nav-frontend-typografi';
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import { ReactComponent as Lukk } from '../../fellesikoner/lukk.svg';
 import { DialogData, StringOrNull } from '../../utils/Typer';
 import { AktivitetskortPreview } from '../aktivitet/AktivitetskortPreview';
 import { AktivitetContextType, useAktivitetContext } from '../AktivitetProvider';
 import styles from './DialogHeader.module.less';
 import DialogHeaderFeil from './DialogHeaderFeil';
+import { DialogOverskrift } from './DialogOverskrift';
 
 export const dialogHeaderID1 = 'dialog_header_1';
 export const dialogHeaderID2 = 'dialog_header_2';
@@ -19,22 +18,6 @@ interface DialogHeaderProps {
     dialog?: DialogData;
     aktivitetId?: StringOrNull;
     visSkygge?: boolean;
-}
-
-interface DialogOverskriftProps {
-    tekst?: StringOrNull;
-}
-
-function DialogOverskrift(props: DialogOverskriftProps) {
-    if (!props.tekst) {
-        return null;
-    }
-
-    return (
-        <Undertittel id={dialogHeaderID2} className={styles.tittel}>
-            {props.tekst}
-        </Undertittel>
-    );
 }
 
 function harAktivitetDataFeil(aktivitetData: AktivitetContextType, aktivitetId?: string | null) {
@@ -51,53 +34,33 @@ export function DialogHeader(props: DialogHeaderProps) {
     const { dialog, aktivitetId, visSkygge } = props;
     const aktivitet = aktivitetId || dialog?.aktivitetId;
 
-    const aktvitetData = useAktivitetContext();
+    const aktivitetData = useAktivitetContext();
+    const erFeil = harAktivitetDataFeil(aktivitetData, aktivitet);
+    const viseAktivitet = aktivitet && !erFeil;
 
-    const erFeil = harAktivitetDataFeil(aktvitetData, aktivitet);
+    const TilbakeKnapp = () => (
+        <Link to="/" title="Til dialoger" className={styles.tilbakeTilOversikt}>
+            <VenstreChevron stor />
+        </Link>
+    );
+
+    const HeaderInnhold = () =>
+        viseAktivitet ? (
+            <AktivitetskortPreview aktivitetId={aktivitet} />
+        ) : (
+            <DialogOverskrift tekst={dialog?.overskrift} />
+        );
 
     return (
         <>
             <DialogHeaderFeil visible={erFeil} />
-            <Header visSkygge={visSkygge}>
+            <div className={classNames(styles.dialogHeader, { [styles.dialogHeaderShadow]: visSkygge })}>
+                <TilbakeKnapp />
                 <div id={dialogHeaderID1} className="hide">
                     Dialog om:
                 </div>
-                {aktivitet && !erFeil ? (
-                    <AktivitetskortPreview aktivitetId={aktivitet} />
-                ) : (
-                    <DialogOverskrift tekst={dialog?.overskrift} />
-                )}
-            </Header>
+                <HeaderInnhold />
+            </div>
         </>
-    );
-}
-
-function Header(props: { children?: React.ReactNode; visSkygge?: boolean; className?: string }) {
-    const { children, visSkygge, className } = props;
-    return (
-        <div className={styles.dialogHeaderContainer}>
-            <div className={classNames(styles.dialogHeader, { [styles.dialogHeaderShadow]: visSkygge }, className)}>
-                <Link to="/" title="Til dialoger" className={styles.tilbakeTilOversikt}>
-                    <VenstreChevron stor />
-                </Link>
-                <div className={styles.headerContent}>{children}</div>
-            </div>
-        </div>
-    );
-}
-
-export function TittelHeader(props: { children?: string }) {
-    return (
-        <div className={classNames(styles.dialogHeader, styles.tittelHeader)}>
-            <Link to="/" title="Til dialoger" className={styles.tilbakeTilOversikt}>
-                <VenstreChevron stor />
-            </Link>
-            <div className={styles.headerContent}>
-                <Systemtittel className={styles.tittel}>{props.children ?? ''}</Systemtittel>
-            </div>
-            <Link to="/" title="lukk" className={styles.lukkeknapp}>
-                <Lukk />
-            </Link>
-        </div>
     );
 }
